rename interface params

+++ tieredstore/src/main/java/org/apache/rocketmq/tieredstore/metadata/TopicMetadata.java
@@ -47,7 +47,7 @@ public class TopicMetadata {
         return topicId;
     }
 
-    public void setTopicId(int topicId) {
+    public void setTopicId(long topicId) {
         this.topicId = topicId;
     }
